Separate the components
