+++ app/components/EventList.vue
